Add migration test
