+++ django_project/stakeholder/tests/test_stakeholder_models.py
@@ -12,7 +12,6 @@ from stakeholder.factories import (
     organisationRepresentativeFactory,
     organisationUserFactory,
     userLoginFactory,
-    userProfileFactory,
     userRoleTypeFactory,
     userTitleFactory,
 )
